Resolve "skoori topelt kontroll"

Closes #65

See merge request sesuur/iti0200-2020-project!83

+++ src/Game/player/Player.java
@@ -206,7 +206,9 @@ public abstract class Player extends ImageView {
     }
 
     public void dropPickedUpFlag() {
-        pickedUpFlag.drop();
+        if (pickedUpFlag != null) {
+            pickedUpFlag.drop();
+        }
         this.pickedUpFlag = null;
     }
